+++ provider.py
@@ -30,6 +30,7 @@ from .algorithms.alg_density_grid import DensityGridAlgorithm
 from .algorithms.alg_performance_report import PlanPerformanceReportAlgorithm
 from .algorithms.alg_facility_location import FacilityLocationAlgorithm
 from .algorithms.alg_capacitated_allocation import CapacitatedAllocationAlgorithm
+from .algorithms.alg_land_allocation import LandUseAllocationAlgorithm
 from .algorithms.alg_access_equity import AccessEquityAlgorithm
 
 
@@ -82,5 +83,6 @@ class PlanXProvider(QgsProcessingProvider):
         # 8 | Optimization
         self.addAlgorithm(FacilityLocationAlgorithm())
         self.addAlgorithm(CapacitatedAllocationAlgorithm())
+        self.addAlgorithm(LandUseAllocationAlgorithm())
         # 9 | Equity
         self.addAlgorithm(AccessEquityAlgorithm())